rename graph.ts to subgraph.ts and move to util
