+++ packages/token-bridge-sdk/src/withdrawals/fetchETHWithdrawalsFromSubgraph.ts
@@ -2,7 +2,7 @@ import { gql } from '@apollo/client'
 import { BigNumber } from '@ethersproject/bignumber'
 import { Provider } from '@ethersproject/providers'
 
-import { getL2SubgraphClient } from './graph'
+import { getL2SubgraphClient } from '../util/subgraph'
 import { L2ToL1EventResult } from '../hooks/arbTokenBridge.types'
 
 /**
